SISRP-45434 - Post-3/3 React Cleanup - move Containers to same directory as Components
