+++ src/react/components/PnpCalculator/CurrentGradePercentage.js
