+++ src/react/components/PnpCalculator/PnpCalculator.js
