+++ src/react/components/PnpCalculator/PnpCalculatorContainer.js
@@ -2,10 +2,10 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import _ from 'lodash';
 import { react2angular } from 'react2angular';
-import { updateStateProperty } from '../helpers/state';
+import { updateStateProperty } from '../../helpers/state';
 
-import RedTimesCircle from '../components/base/icon/RedTimesCircle';
-import PnpCalculator from '../components/pnp_calculator/PnpCalculator';
+import RedTimesCircle from '../base/icon/RedTimesCircle';
+import PnpCalculator from './PnpCalculator';
 
 const propTypes = {
   $scope: PropTypes.object.isRequired,
