+++ src/react/components/PnpCalculator/RatioCalculation.js
