+++ src/react/components/base/link/CampusSolutionsLinkContainer.js
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import _ from 'lodash';
 
-import { fixLastQuestionMark, updateQueryStringParameter } from '../helpers/links';
-import CampusSolutionsLink from '../components/base/link/CampusSolutionsLink';
-import CampusSolutionsOutboundLink from '../components/base/link/CampusSolutionsOutboundLink';
+import { fixLastQuestionMark, updateQueryStringParameter } from '../../../helpers/links';
+import CampusSolutionsLink from './CampusSolutionsLink';
+import CampusSolutionsOutboundLink from './CampusSolutionsOutboundLink';
 
 const propTypes = {
   linkObj: PropTypes.object.isRequired
